+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 development:
   * prevent tux to be required
   * add sinatra assets pipeline extension
+  * restructure generated app
 
 0.4.0:
   * add SQL logger to development environment