+++ CHANGELOG.md
@@ -5,6 +5,7 @@ development:
   * use erb for app templates
   * create Rakefile generator with and without database
   * add new Sequel configuration when generating app with database
+  * add Sinatra::Cache extension
 
 0.4.0:
   * add SQL logger to development environment